Solved 2023 05_1

+++ 2023/Solver.cs
@@ -3,6 +3,6 @@ namespace AdventOfCode2023;
 abstract class Solver
 {
     public abstract string PuzzleInputPath { get; }
-    public abstract int SolvePart1();
-    public abstract int SolvePart2();
+    public abstract long SolvePart1();
+    public abstract long SolvePart2();
 }
