+++ 2023/Solvers/Day03Solver.cs
@@ -5,7 +5,7 @@ class Day03Solver : Solver
     private string _puzzleInputPath = "PuzzleInputs/03.txt";
     public override string PuzzleInputPath { get => _puzzleInputPath; }
 
-    public override int SolvePart1()
+    public override long SolvePart1()
     {
         string[] lines = File.ReadAllLines(PuzzleInputPath);
         int count = 0;
@@ -37,7 +37,7 @@ class Day03Solver : Solver
         return count;
     }
 
-    public override int SolvePart2()
+    public override long SolvePart2()
     {
         string[] lines = File.ReadAllLines(PuzzleInputPath);
         int count = 0;
